# internal/presentation/openapi/generate/types.gen.go
// Package openapi provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.0 DO NOT EDIT.
package openapi

import (
	"time"
)

// Defines values for ErrorCode.
const (
	InternalServerError ErrorCode = "internal_server_error"
)

// Article defines model for Article.
type Article struct {
	// Id 記事ID（UUID形式）
	Id string `json:"id"`

	// PublishedAt 公開日時
	PublishedAt time.Time `json:"published_at"`
	Tags        []string  `json:"tags"`

	// Title 記事タイトル
	Title string `json:"title"`
}

// ArticlesGetResponseBody defines model for ArticlesGetResponseBody.
type ArticlesGetResponseBody struct {
	Articles []Article `json:"articles"`
}

// Error defines model for Error.
type Error struct {
	// Code エラーコード
	Code ErrorCode `json:"code"`

	// Details エラーの詳細情報
	Details *map[string]string `json:"details,omitempty"`

	// Message エラーメッセージ
	Message string `json:"message"`
}

// ErrorCode エラーコード
type ErrorCode string

// RequestID リクエストID（UUID形式）
type RequestID = string

// ArticlesGetParams defines parameters for ArticlesGet.
type ArticlesGetParams struct {
	// PublishYear 公開年でフィルタリング
	PublishYear *int `form:"publish-year,omitempty" json:"publish-year,omitempty"`

	// Tags タグでフィルタリング（複数指定可能）
	Tags *[]string `form:"tags,omitempty" json:"tags,omitempty"`
}